Ajout d'un nouveau event pour remettres les hélico dans le garages
Ajout d'un nouveau event, qui fait que a chaque redémarage du serveur
les hélicoptère sont dans le garages
