+++ __resource.lua
@@ -1,10 +1,11 @@
 server_scripts {
-	'config.lua',
-	'server.lua',
-	'helishop_s.lua'
+	'helishop/config.lua',
+	'helishop/server.lua',
+	'helishop/helishop_s.lua',
+	'shop_postload/server.lua'
 }
 client_script {
-	'client.lua',
-	'GUI.lua',
-	'helishop.lua'
+	'helishop/client.lua',
+	'helishop/GUI.lua',
+	'helishop/helishop.lua'
 }
